refactor: migrate PageLoading and PageError components to Utils for improved organization and reuse across admin routes
