+++ web/src/platform/routes/admin/Operations.tsx
@@ -9,7 +9,7 @@ import { api } from '@/lib/api';
 import { dateTimeFormatter } from '@/lib/utils';
 import { usePaginate } from '@/lib/hooks/pagination';
 import { Table, TableColumn } from '@/components/ui/Table';
-import { PageError, PageLoading } from '@/components/layouts/State';
+import { PageError, PageLoading } from '@/components/Utils';
 import { zOperationResponse } from '@/lib/generated/platform-api-v1/zod.gen';
 
 /** Renders the admin operations page. */
