+++ web/src/platform/routes/admin/Storage.tsx
@@ -13,8 +13,8 @@ import { useDeleteDialog } from '@/lib/utils';
 import { useToast } from '@/lib/hooks/use-toast';
 import { usePaginate } from '@/lib/hooks/pagination';
 import { Table, TableColumn } from '@/components/ui/Table';
+import { PageError, PageLoading } from '@/components/Utils';
 import CreateStorage from '@/components/dialogs/CreateStorage';
-import { PageError, PageLoading } from '@/components/layouts/State';
 import { DeleteConfirmation } from '@/components/dialogs/DeleteConfirmation';
 import { zStorageRegistryResponse } from '@/lib/generated/platform-api-v1/zod.gen';
