+++ web/src/platform/routes/admin/Users.tsx
@@ -13,7 +13,7 @@ import { api } from '@/lib/api';
 import { useToast } from '@/lib/hooks/use-toast';
 import { usePaginate } from '@/lib/hooks/pagination';
 import { Table, TableColumn } from '@/components/ui/Table';
-import { PageError, PageLoading } from '@/components/layouts/State';
+import { PageError, PageLoading } from '@/components/Utils';
 import { zUserSummary } from '@/lib/generated/platform-api-v1/zod.gen';
 
 /** Renders the admin users page. */
